Add lint command.
This command reads and re-writes (in-place) the metadata files.
